requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
pyserial==3.5
requests==2.31.0
Pillow==10.0.1
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.2
seaborn==0.12.2
openpyxl==3.1.2
reportlab==4.0.4
python-barcode==0.15.1
qrcode==7.4.2
cryptography==41.0.4
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==2.0.0
phonenumbers==8.13.17
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
celery==5.3.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
pymysql==1.1.0
sqlite3-to-mysql==1.4.16
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.1.1
flask-mail==0.9.1
flask-cors==4.0.0
flask-restful==0.3.10
marshmallow==3.20.1
webargs==8.3.0
apispec==6.3.0
swagger-ui-bundle==0.0.9
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.5.1
pre-commit==3.4.0
